+++ changelog.md
@@ -11,6 +11,7 @@
 -   requesting undefined IDs due to incorrect format
 -   map PP all values being the same
 -   fix some conversions
+-   score stats being mutated by calc FC function
 
 ### Changed